// src/app/shared/constants/chart-types.enum.ts
export enum ChartTypes {
    ARC_TOTAL = 'ARC_TOTAL',
    ARC_ONLINE = 'ARC_ONLINE',
    ARC_OFFLINE = 'ARC_OFFLINE',
    ACTION_ITEM_STATUS = 'ACTION_ITEM_STATUS',
    ACTION_ITEM_AGEING = 'ACTION_ITEM_AGEING',
    DEVIATION_ITEM_STATUS = 'DEVIATION_ITEM_STATUS',
    DEVIATION_ITEM_AGEING = 'DEVIATION_ITEM_AGEING',
    COE_ITEM_STATUS = 'COE_ITEM_STATUS',
    COE_ITEM_AGEING = 'COE_ITEM_AGEING'
}